requirements.txt
# Core ML / embeddings
FlagEmbedding>=1.3.5
torch>=2.11.0          # install the CUDA build separately — see README

# LlamaIndex
llama-index-core>=0.14.20
llama-index-llms-ollama>=0.10.1
llama-index-readers-file>=0.6.0

# Vector database
qdrant-client>=1.17.1

# API server
fastapi>=0.135.3
uvicorn>=0.44.0

# HTTP client (chat.py)
httpx>=0.28.1

# HTML parsing (ingest.py)
beautifulsoup4>=4.14.3
